python/requirements.txt
protobuf==3.20.0
python-snappy
requests
